requirements.txt.txt
streamlit==1.50.0
google-genai==1.45.0
python-dotenv==1.1.1
